Add liberate test and build_test invoke tasks (bug 1365907)
Links are created to allow updating the phutil_map file.
As file volumes in docker require to be provided with an absolute path,
/tmp/phutil_map is used. It needs to be linked to phutil_map to update
it directly from `arc liberate` command.

For the `arc unit` tests to run there is a need to have database
running. The phabdb from conduit-demo was used to create a minimum
viable container.