Run tests
  - circle.yml: now that the tasks for the tests exist, let's make circle actually
run them
  - tasks.py: added linefeeds to make flake8 shut up. ;), Docstring too.
  - circle.yml: work around broken circleci docker-compose
  - tasks.py: debug the link step
  - circle's version of docker-compose doesn't support the version "2"
syntax in docker-compose.test.yml. Using a hint found here:
https://github.com/circleci/docker/issues/5, we are able to load a
newer version of docker-compose into a location earlier in the path that
allows the build to complete with a newer version of docker-compose.

Fix errors coming from removing btrfs root volumes in docker builds

  - tasks.py: attempting to fix https://github.com/moby/moby/issues/22547 via https://circleci.com/docs/1.0/docker-btrfs-error/; explicitly set rm=false
  - tasks.py: eventually, remove the rm completely to fix https://github.com/moby/moby/issues/22547; successful fix
  - rather than linking the file, it makes more sense (and works better)
to copy the file into place where the fs expects it to be. All attempts
to do this with linking fail, cp /tmp works.

Remove debugging statements.