Refactored workflow for phabutil map
Creating a link on the host machine was a weak spot. Maris came out with
a different idea, so we could avoid that. `phutil_map` is moved to a
subdirectory and entire subdirectory is used as a volume in container.
Then the `phabutil_map` file on the container is linked to
`/app/phabricator/src/__phutil_library_map__.php`. It is safer and makes
the `tasks.py` cleaner.