Add TEST_PLAN to document how to test the Integration (Bug 1394434). r?davidwalsh
Added a doc that describes the different use cases of Phabricator/BMO integration.
This doc includes example test cases which can be manually tested when changing
the integration to prevent regressions.

Also included at the end is a specific guide on how to get a local test environment
up and running. The tests are general enough that they can be run in a local
environment or in the development, staging, or (if needed) production.

This is meant to be a living document so as we finish the integration more tests
will be added, and as functionality changes this doc should as well.