WIP: Bug 1397927 - Add comment to bug when a revision is accepted/denied with the users real names
Summary:
Add comment to bugzilla bug when a revision is accepted or rejected that shows who the acceptor is and a link back to the transaction in revision.

These changes in Doorkeeper are to support changes on the BMO sides that use the rejected bmo user ids to add comment to the bug.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Bug #: 1397927

Differential Revision: https://phabricator.services.mozilla.com/D41