Bug 1398891 - Allow bug requirement to be toggled
Summary: This adds a config var for turning on or off the bug id requirement.

Test Plan: Submit revisions with the config value off and on.  Confirm the revision submission is or isn't stopped as with or without a bug value.  Monitor the Doorkeeper daemon to ensure no syncing occurs when no bug number, and that syncing continues working with the bug ID provided.

Reviewers: dkl

Reviewed By: dkl

Subscribers: davidwalsh

Bug #: 1398891

Differential Revision: https://phabricator.services.mozilla.com/D52