Bug 1397454 - Use Doorkeeper to ping BMO when a new revision is created to update view policy
Summary:
Currently we rely on a combination of a Herald Rule and a Harbormaster Build Plan to ping BMO
when a new revision is created. Then BMO will use the revision ID to update the revisions view/edit
policies appropriately. This patch is about moving away from the complicated setup and use the
Doorkeeper extension code to do the same.

Test Plan: None

Reviewers: davidwalsh

Reviewed By: davidwalsh

Bug #: 1397454

Differential Revision: https://phabricator.services.mozilla.com/D35