Fix 1400026 - Customize retry intervals and attempts.  Also render detail for works screen.
Summary: Customizes a few of the items we spoke about in today's meeting.

Test Plan: Add an exception anywhere in the `publishFeedStory` method to purposely bomb an task so that it will be re-queued.  Go to the /daemon/ address, while logged in as admin, and see the task listed under Leased Tasks -- you should see the Failures count increase as you refresh.  If you care to run your own daemon, leave it open and watch the failures continue to happen.  Next, click the "View" button next to a failed task and you should see a new detail item, Data, which provides a link to the bombing story's revision.

Reviewers: dkl

Reviewed By: dkl

Bug #: 1400026

Differential Revision: https://phabricator.services.mozilla.com/D59