Fix bug 1401748 - Make bugless revisions public when created or when bug ID is removed
Summary:
Makes a bug public when:

No bug is provided upon initial revision creation
A bug ID was associated with the revision but is then removed.
In the case #2, the revision should be obsoleted on the original bug.

Test Plan:
Try these:

Create a new revision with no bug ID and with a view or edit policy of admins. Refresh the page. The revision should be public. (You have to refresh the page because, in the mean time, doorkeeper is doing its thing to make the revision public)
Create a new revision with a bug ID. View the attachment on BMO. Remove the bug ID from phab. The BMO attachment should be obsoleted.
Create a new revision without a bug ID. View no attachment on BMO. Add a bug ID. View the attachment on BMO. Remove the bug ID. The BMO attachment should be deleted.
Test another functionality that would call for a bug to be obsoleted, ensure that it works.
Any other crazy assortment we can think of.

Reviewers: dkl

Reviewed By: dkl

Subscribers: davidwalsh

Bug #: 1401748

Differential Revision: https://phabricator.services.mozilla.com/D74