Make extensions a separate library.
Summary:
Currently we're modifying the Phabricator library by adding extension
files to the main phutil library map. This patch follows the documentation
[1], [2]. It makes the extensions a custom library and configures the
Phabricator to load it.

Extensions files are moved to `extensions/src` directory.

PRODUCTION
`extensions/` directory is copied to `/app/moz-extensions/` in Dockerfile.
Custom Phabricator configuration is created in Dockerfile so production
can load moz-extensions library. It is copying the
`moz-extensions.conf.php` file to `conf/custom/`.

`entrypoint.sh` makes calls to `bin/config` to make configuration changes
we need at startup in development, but these calls don't actually
happen in production. If we wanted to use the same method to add
our extension library to load-libraries [3] we'd be overwriting the
configuration file provided by ops, which would be bad.
`conf/local/ENVIRONMENT` isn't touched by operations though
so we are safe to add our load-libraries configuration that way.

DEVELOPMENT
Arcanist needs to load the extensions library so it can be tested.
We need to specify  the phutil library position in `.arcconfig`.
`test-arcconfig` file is created for that task and copied to the container
in `Dockerfile-test`.

`extensions/` directory is configured as a volume in test container so
created phabutil map is updated in the repository.

Nits fixed:
Dockerfile - split lines in patching command
Dockerfile-test - nicer `apk add` command
Readme.md - when to use `arc liberate` info updated

[1] https://secure.phabricator.com/book/phabcontrib/article/adding_new_classes/
[2] https://secure.phabricator.com/book/phabricator/article/advanced_configuration/#creating-a-configuration
[3] `./bin/config set load-libraries '{"moz-extensions": "/app/moz-extensions/src"}'`

Test Plan:
invoke test
Please check all of the extensions using bmo-extensions repository.

Reviewers: dkl, smacleod

Reviewed By: dkl, smacleod

Subscribers: mcote

Bug #: 1411222

Differential Revision: https://phabricator.services.mozilla.com/D157