conduit api: Create new feed.query_id conduit API method to page using the id column instead of the chronologicalKey value.
Summary:
After some thorough discussion, it was decided to that it would be better to use the ID column value for paging transactions in
the custom API endpoint feed.query_id due to the method of how chronologicalKey (default paging value) is generated.
chronologicalKey is a combination of the epoch and a random generated value. It was determined that it could be possible
to have transactions out of order or missed entirely by using it for paging and filtering of transactions. Missing transactions
when using feed.query could be detrimental to how BMO uses the transaction data to determine revision privacy and other
action due to revision changes.

Changes:

- Create new feed.query_id Conduit API call to allow paging based in story id instead of chronologicalKey.
- Patch to Phabricator code to remove final/protected declarations for classes needed for above.
- Subclassed PhabricatorQueryFeed to make id the default paging column instead of chronologicalKey.
- Updated phutil_library_map to include new subclassed libraries.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Bug #: 1418004

Differential Revision: https://phabricator.services.mozilla.com/D259