Bug 1428791 - Port PolicyQueryConduiitAPIMethod from wikimedia to our Phabricator instance
Summary: This method will allow BMO to check to see if a new custom policy matches the one current set on a revision and skip modification if they are the same.

Test Plan:
Create revision with custom policy and obtain the PHID for the policy.
User policy.query API method to look up information about the policy.

Reviewers: davidwalsh

Reviewed By: davidwalsh

Bug #: 1428791

Differential Revision: https://phabricator.services.mozilla.com/D357