Bug 1438004: Using extension code, update user.search to return external account information such as user id
Summary: Currently Bugzilla has to make an additional API call to the Conduit API to map PHIDs to Bugzilla IDs when it wants to find the user information in the Bugzilla database. Adding the Bugzilla account IDs to the results returned by the user.search API method will remove the need for the extra API call. Bugzilla will just need to request the 'external-accounts' attachment extension when using user.search.

Test Plan:
To test:

Construct an API call to the user.search method either using 'arc conduit' or the UI form at http://phabricator.test/api/user.search.

The JSON query parameters should look similar to:

{
  "constraints":{
    "phids": ["PHID-USER-h4aqihzqsnytz7nsegsr"]
  },
  "attachments":{
    "external-accounts":1
  }
}

In the results. You should see something like:

{
   "data": [
     {
       "id": 2,
       "type": "USER",
       "phid": "PHID-USER-h4aqihzqsnytz7nsegsr",
       "fields": {
         "username": "phab-bot",
         "realName": "Phabricator Automation",
         "roles": [
           "admin",
           "verified",
           "approved",
           "activated"
         ],
         "dateCreated": 1512573120,
         "dateModified": 1512574523,
         "policy": {
           "view": "public",
           "edit": "no-one"
         }
       },
       "attachments": {
         "external-accounts": {
           "external-accounts": [
             {
               "id": "9",
               "type": "bmo"
             }
           ]
         }
       }
     }
   ],

The PHID you use in the constraints may be different and would need to be adjusted. You can also leave off the constraints and just see a list of all users in the database.

Reviewers: zalun

Reviewed By: zalun

Bug #: 1438004

Differential Revision: https://phabricator.services.mozilla.com/D592