lando integration: Add [View in Lando] link.
Summary:
Creates a "View in Lando" link on the Revision page.
Constructs the URI using `lando-ui.url` custom config, the Revision ID and the
active Diff ID.

If the `lando-ui.url` config is not set the link is not rendered.

WARNING:
This patch removes the default value of a config. Please check if it has the
right value before deployment.

Test Plan:
Observe 3 states of the link:
1. Not displayed - lando-ui.url is not set
2. Disabled (greyed out) - Revision is not accepted
3. Enabled - Revision is accepted

Reviewers: dkl, smacleod

Reviewed By: dkl, smacleod

Bug #: 1443510

Differential Revision: https://phabricator.services.mozilla.com/D715