Bug 1444172 - Provide better structured error logging from Phabricator that is easy to monitor
Summary:
We need to either extend the current MozLogging extension that we have installed to include all
errors that Phab generates or add in support for logging to external tools such Sentry, etc.
Currently we only output mozlog formatted messages for the BMO auth adaptor. We would need to
latch in somewhere to be able to grab all error messages and output them to mozlog log file.

Test Plan:
1. Configure Phabricator to use the Sentry DSN provided by to you.
2. Perform actions in Phabricator that will cause exceptions/errors such as submitting a non-patch for review.
3. Observe the exceptions being recorded in https://sentry.prod.mozaws.net/operations/mozphab-dev

Reviewers: zalun

Reviewed By: zalun

Bug #: 1444172

Differential Revision: https://phabricator.services.mozilla.com/D776