Bug 1445866 - email not being sent for review requests
Summary: Create new PolicyRule to allow a custom policy to also apply reviewers to the allow list

Test Plan:
Create custom policy that includes one or more projects, allow subscribers, and allow reviewers.
Add a reviewer to the revision that is not a subscriber or member of any of the secure groups.
Observe that the reviewer can access the private revision and also see it in their home dashboard.

Reviewers: zalun

Reviewed By: zalun

Subscribers: mcote, glob

Bug #: 1445866

Differential Revision: https://phabricator.services.mozilla.com/D792