secure revision: Add warning about the workflow for secure revisions.
Summary:
Secure revisions are managed in Bugzilla. Users should be notified about
this behaviour.

To place a warning info on the revision's page a new plugin is created.
Phabricator's code is patched so it will call the plugin and place it
above the "Details" panel and any potential warnings from Phabricator.

Plugin is extendable - we can add warnings, to the page by modifying the
plugin only.

{F417393}

Test Plan: Open revision page with no, secure-revision and other project assigned

Reviewers: dkl

Reviewed By: dkl

Bug #: 1446973

Differential Revision: https://phabricator.services.mozilla.com/D795