cinnabar: Abort when local VCS is different than the repository one.
Summary:
Creating a diff will now throw an exception if local VCS is not the same
as defined in the repository on Phabricator.

Test Plan:
Using demo environment

Using git to commit to VCT repo is not allowed
- create a Mercurial repository in Phabricator (callsign VCT)
- create a git repository in demo image
- copy ../version-control-tools/.arcanist, commit as initial
- create another commit
- try to create a diff using arc
Observe it failing.

Proceed normally to create a diff using hg in version-control-tools dir

Reviewers: smacleod, zalun

Bug #: 1458191

Differential Revision: https://phabricator.services.mozilla.com/D1131