phabext: Add Mozilla MOTD CSS File
Summary:
MOTD info window shows up under the resize bar if file tree is
switched on. This patch adds a CSS file which is fixing this issue.

**WARNING:**
There might be a need to do an administrative task:

If an Exception of missing celerity resource is thrown - run
$ docker-compose exec phabricator sh
$ ./phabricator/bin/celery map

The exception is nondeterministic. I've seen it once in the suite on a docker
image created in `phabricator-extensions` dir. Later I've removed all images
and rerun the suite, but the Exception was not there anymore. I've never
observed it again.
Please keep an eye on this to avoid the problem in dev/production.

Test Plan: Please report if you've seen the above exception.

Reviewers: dkl

Reviewed By: dkl

Subscribers: glob

Bug #: 1475019

Differential Revision: https://phabricator.services.mozilla.com/D2184