Bug 1488268 - Not all open review requests are visible on phabricator's dashboard
Summary:
The home page code limits the total number of revisions in the active revisions panel to 15
and then those are divided into their respective categories. This is causing many people to
miss some of their reviews if they do not know how to show the full list. This patch raises
the limit to 500.

Test Plan: See more than 15 revisions on the default home page.

Reviewers: zalun

Reviewed By: zalun

Bug #: 1488268

Differential Revision: https://phabricator.services.mozilla.com/D5148