Bug 1490795 - Decrease z-index for the info view. r=dkl
Summary:
Info view is used for the MOTD and informations on the dashboard.
It has `z-index` set to `6` to display the MOTD box over the
`.phabricator-nav-drag` element. The notification drop-down
(`.phabricator-notification-menu`) is then displayed below MOTD and other
info view elements.

Fortunately the dragging bar element has `z-index` set to `4`. This patch
is setting the `z-index` of the `.phui-info-view` elements to `5`.

Test Plan:
{F601242}

{F601241}

{F601240}

Reviewers: dkl

Reviewed By: dkl

Bug #: 1490795

Differential Revision: https://phabricator.services.mozilla.com/D11464