differential: copy commit info to diff on commit sync (Bug 1489126).
Summary:
Phabricator does not copy the commit meta data (such as author, commit
message, and id) to the diff when updating a revision from a synced
commit. We have requested this upstream[1] but will patch Phabricator
until it is fixed.

This change does not make Phabricator copy all commit information, only
the most important data that we care about. This might require a
followup as there is also a concerning TODO which indicates changes to
binary files might not be part of the diff that Phabricator updates the
revision with.

[1] https://admin.phacility.com/PHI787

Test Plan:
Iterated on `1489126_1.patch` by modifying the code and running
`docker-compose -f docker-compose.yml -f docker-compose.bmo.yml exec phabricator /app/phabricator/bin/differential extract <commit-sha>`
to create diffs. After that was working verified that a normal
commit sync would have the same result by creating a revision
and closing it by pushing a commit.

Reviewers: dkl

Reviewed By: dkl

Bug #: 1489126

Differential Revision: https://phabricator.services.mozilla.com/D24010