differential: re-open revisions backed out by a commit (Bug 1507578).
Summary:
Differential revisions will now be re-opened by audit when diffusion
imports a backout commit. This feature was requested upstream[1] but
they do not want it in the core phabricator product, so we will use a
patch to make it work.

It is important that the `differential.allow-reopen` config entry be set
to `true` or commit importing will fail and repeatedly retry. This
shouldn't be a problem in Mozilla's production install as that is the
config value we use.

[1] https://admin.phacility.com/PHI1008

Test Plan:
Started phabricator locally and caused it to import a backout commit.
Observed that the revision was re-opened as expected.

Reviewers: dkl

Reviewed By: dkl

Bug #: 1507578

Differential Revision: https://phabricator.services.mozilla.com/D27758