phabricator-extensions: Bug 1596023 "New Changes" link for differentials
A new event listener was added to put the "New Changes" link on the revision page.

This link is disabled for users who are already viewing "New Changes", as well as for unauthenticated users.
I'd prefer to add some hover-text for the link when it's disabled, but I couldn't immediately see how.
I'll leave that work for the follow up ticket 1604841.

Differential Revision: https://phabricator.services.mozilla.com/D57720