Bug 1671677: Titles should not include newlines
Titles should never have newlines - moz-phab moves everything past the
first newline into the patch description. However, if an alternative
patch submission tool is used, then newlines can still occur.

This triggers an error case on the email-side, since the title is used
in the email subject, and email subjects cannot have newlines.
So, the solution here is to replace the erroneous newlines with spaces.

Differential Revision: https://phabricator.services.mozilla.com/D93871