Bug 1671984: Don't create "/app" volume
The "/app" volume is inhibiting local development, forcing volumes to
be dropped in order to make changes to files such as the entrypoint.
Since the volume was only used for sharing "site.conf" in the local
development environment, we move "site.conf" to a separate directory
which will be volume-ized and shared instead.

Differential Revision: https://phabricator.services.mozilla.com/D97022