Bug 1672239: Add "minimalContext" to each email event
Adjusts the structure of the output so that the consuming service can
optionally use a "minimal context" to produce emails if the full context
can't be rendered.

Changes exception-handling behaviour:
* If an error is encountered while resolving email context, safely fall
back to just providing a "minimal context".
* If an error is encountered before the "minimal context" can be
resolved, **no longer skip** the event. Instead, respond with an error,
and block the email queue until the issue is resolved.

Differential Revision: https://phabricator.services.mozilla.com/D111181