Bug 1669586: Change EventKind workflow to be more obvious
Rather than iterating over the transactions and piecing together
what they represent, instead we can target individual events: "this
is $ABANDON if there is an expected type in the list of transactions".

The benefit here is most obvious when comparing how we detect
a "$CREATE" event - rather than putting together the data
at the end of iteration, the logic is now centralized.

Depends on D115388

Differential Revision: https://phabricator.services.mozilla.com/D114399