Bug 1717515: Add patch to avoid notif backlog when using Moz emails
This patch is a little bit risky, but:
* Phabricator is seeing less upstream changes, which lowers risk, and
* I opted this patch to have as little impact as possible.

It works by adding a new "deliverable" state to mail actors: in addition
to being either "deliverable" (by built-in emails) or "undeliverable"
(not receiving any emails), there's now a third state:
"externally-deliverable". This allows existing Phabricator logic around
sending emails to continue seeing Moz-Email users as "undeliverable",
while also allowing one use-case - the notification panel - to see such
users as "not _un_deliverable".

I deliberately renamed `filterDeliverableActors()` so that, if changes
upstream cause it to gain a new caller, the patch will fail to apply.
If/when that happens, we'll need to consider how the new caller should
behave around Moz-Emails, and adjust the patch accordingly.

Differential Revision: https://phabricator.services.mozilla.com/D119537